requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
pycodestyle==2.7.0
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
